Refactor files & codes as my preferences
