+++ app.js
@@ -4,8 +4,8 @@ const express = require('express');
 const app = express();
 
 // error handler
-const notFoundMiddleware = require('./middleware/not-found');
-const errorHandlerMiddleware = require('./middleware/error-handler');
+const notFoundMiddleware = require('./middleware/notFoundMiddleware');
+const errorHandlerMiddleware = require('./middleware/errorHandlerMiddleware');
 
 app.use(express.json());
 // extra packages
