+++ errors/CustomError.js
@@ -1,0 +1,7 @@
+class CustomError extends Error {
+  constructor(message) {
+    super(message)
+  }
+}
+
+module.exports = CustomError
