--- errors/custom-api.js
@@ -1,7 +1,0 @@
-class CustomAPIError extends Error {
-  constructor(message) {
-    super(message)
-  }
-}
-
-module.exports = CustomAPIError
