--- middleware/not-found.js
@@ -1,3 +1,0 @@
-const notFound = (req, res) => res.status(404).send('Route does not exist')
-
-module.exports = notFound
